+++ lang/src/idl.rs
@@ -17,8 +17,6 @@
 //! Note that IDL account instructions are automatically inserted into all
 //! Anchor programs. To remove them, one can use the `no-idl` feature.
 
-#[allow(deprecated)]
-use crate::accounts::program_account::ProgramAccount;
 use crate::prelude::*;
 use solana_program::pubkey::Pubkey;
 
@@ -35,82 +33,21 @@ pub const IDL_IX_TAG_LE: [u8; 8] = IDL_IX_TAG.to_le_bytes();
 pub const ERASED_AUTHORITY: Pubkey = Pubkey::new_from_array([0u8; 32]);
 
 #[derive(AnchorSerialize, AnchorDeserialize)]
-pub enum IdlInstruction {
-    // One time initializer for creating the program's idl account.
-    Create { data_len: u64 },
-    // Creates a new IDL account buffer. Can be called several times.
-    CreateBuffer,
-    // Appends the given data to the end of the idl account buffer.
-    Write { data: Vec<u8> },
-    // Sets a new data buffer for the IdlAccount.
-    SetBuffer,
-    // Sets a new authority on the IdlAccount.
-    SetAuthority { new_authority: Pubkey },
-    Close,
-    // Increases account size for accounts that need over 10kb.
-    Resize { data_len: u64 },
-}
-
-// Accounts for the Create instruction.
-pub type IdlCreateAccounts<'info> = crate::ctor::Ctor<'info>;
-
-// Accounts for Idl instructions.
-#[derive(Accounts)]
-pub struct IdlAccounts<'info> {
-    #[account(mut, has_one = authority)]
-    #[allow(deprecated)]
-    pub idl: ProgramAccount<'info, IdlAccount>,
-    #[account(constraint = authority.key != &ERASED_AUTHORITY)]
-    pub authority: Signer<'info>,
-}
-
-// Accounts for resize account instruction
-#[derive(Accounts)]
-pub struct IdlResizeAccount<'info> {
-    #[account(mut, has_one = authority)]
-    #[allow(deprecated)]
-    pub idl: ProgramAccount<'info, IdlAccount>,
-    #[account(mut, constraint = authority.key != &ERASED_AUTHORITY)]
-    pub authority: Signer<'info>,
-    pub system_program: Program<'info, System>,
-}
-
-// Accounts for creating an idl buffer.
-#[derive(Accounts)]
-pub struct IdlCreateBuffer<'info> {
-    #[account(zero)]
-    #[allow(deprecated)]
-    pub buffer: ProgramAccount<'info, IdlAccount>,
-    #[account(constraint = authority.key != &ERASED_AUTHORITY)]
-    pub authority: Signer<'info>,
-}
-
-// Accounts for upgrading the canonical IdlAccount with the buffer.
-#[derive(Accounts)]
-pub struct IdlSetBuffer<'info> {
-    // The buffer with the new idl data.
-    #[account(mut, constraint = buffer.authority == idl.authority)]
-    #[allow(deprecated)]
-    pub buffer: ProgramAccount<'info, IdlAccount>,
-    // The idl account to be updated with the buffer's data.
-    #[account(mut, has_one = authority)]
-    #[allow(deprecated)]
-    pub idl: ProgramAccount<'info, IdlAccount>,
-    #[account(constraint = authority.key != &ERASED_AUTHORITY)]
-    pub authority: Signer<'info>,
-}
-
-// Accounts for closing the canonical Idl buffer.
-#[derive(Accounts)]
-pub struct IdlCloseAccount<'info> {
-    #[account(mut, has_one = authority, close = sol_destination)]
-    #[allow(deprecated)]
-    pub account: ProgramAccount<'info, IdlAccount>,
-    #[account(constraint = authority.key != &ERASED_AUTHORITY)]
-    pub authority: Signer<'info>,
-    #[account(mut)]
-    pub sol_destination: AccountInfo<'info>,
-}
+    pub enum IdlInstruction {
+        // One time initializer for creating the program's idl account.
+        Create { data_len: u64 },
+        // Creates a new IDL account buffer. Can be called several times.
+        CreateBuffer,
+        // Appends the given data to the end of the idl account buffer.
+        Write { data: Vec<u8> },
+        // Sets a new data buffer for the IdlAccount.
+        SetBuffer,
+        // Sets a new authority on the IdlAccount.
+        SetAuthority { new_authority: Pubkey },
+        Close,
+        // Increases account size for accounts that need over 10kb.
+        Resize { data_len: u64 },
+    }
 
 // The account holding a program's IDL. This is stored on chain so that clients
 // can fetch it and generate a client with nothing but a program's ID.
@@ -135,24 +72,5 @@ impl IdlAccount {
     }
     pub fn seed() -> &'static str {
         "anchor:idl"
-    }
-}
-
-use std::cell::{Ref, RefMut};
-
-pub trait IdlTrailingData<'info> {
-    fn trailing_data(self) -> Ref<'info, [u8]>;
-    fn trailing_data_mut(self) -> RefMut<'info, [u8]>;
-}
-
-#[allow(deprecated)]
-impl<'a, 'info: 'a> IdlTrailingData<'a> for &'a ProgramAccount<'info, IdlAccount> {
-    fn trailing_data(self) -> Ref<'a, [u8]> {
-        let info = self.as_ref();
-        Ref::map(info.try_borrow_data().unwrap(), |d| &d[44..])
-    }
-    fn trailing_data_mut(self) -> RefMut<'a, [u8]> {
-        let info = self.as_ref();
-        RefMut::map(info.try_borrow_mut_data().unwrap(), |d| &mut d[44..])
     }
 }
