+++ lang/syn/src/codegen/program/mod.rs
@@ -8,6 +8,7 @@ mod dispatch;
 mod entry;
 mod handlers;
 mod instruction;
+mod idl;
 
 pub fn generate(program: &Program) -> proc_macro2::TokenStream {
     let mod_name = &program.name;
